% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{attach_location_coordinates}
\alias{attach_location_coordinates}
\title{attach_location_coordinates}
\usage{
attach_location_coordinates(dataframe)
}
\arguments{
\item{data}{rename headers of blups dataframe e.g ag_40 to Ago_owu, ib_12 = Ibadan}
}
\value{
a dataframe with the exact correct names of location
}
\description{
attach_location_coordinates
}
\examples{
# generate_sindex_data(datapath = "path/to/BLUPS.xls")
}
