% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{calculate_selection_index}
\alias{calculate_selection_index}
\title{calculate_selection_index}
\usage{
calculate_selection_index(dataframe, ...)
}
\arguments{
\item{dataframe}{The dataframe without the selection index}
}
\value{
a dataframe with the calculated dataframe
}
\description{
calculate_selection_index
}
\examples{
# Please do ensure that variable argument names matches table headers
# calculate_select_index(dataframe = dat, rtwt = 34, shtwt = 50, dm = 89)
# enter traits as list or vector
# traits = c(shtwt = 6, rtwt = 3, dm = 45, pltht = 23)
# calculate_selection_index(dataframe = dat, traits)


}
