% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindex.R
\name{calculate_sindex_checkmean}
\alias{calculate_sindex_checkmean}
\title{calculate_sindex_checkmean}
\usage{
calculate_sindex_checkmean(dataframe, checks)
}
\arguments{
\item{dataframe}{selection index dataframe to calculate check difference}

\item{checks}{a list of checks in our dataframe to use as a benchmark for our calculation}
}
\value{
a dataframe datatype
}
\description{
calculate_sindex_checkmean
}
\examples{
# calculate_sindex_checkmean(dataframe = "sindex.csv", checks_list)
}
