% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{check_si}
\alias{check_si}
\title{Datatoy}
\usage{
check_si(datapath = "", ...)
}
\arguments{
\item{datapath}{BLUPS data path to check for sindex sheet}
}
\value{
a string datatype, TRUE or FALSE
}
\description{
Datatoy
}
\examples{
# check_si(datapath = "path/to/BLUPS.xls")
}
