% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{distribution_plot}
\alias{distribution_plot}
\title{distribution_plot}
\usage{
distribution_plot(x, y, data, color, lm, text)
}
\arguments{
\item{datapath}{the exact path to your blups data}
}
\value{
a stacked dataframe of all the traits and location in one dataframe
}
\description{
distribution_plot
}
\examples{
# get_genotype_by_location_data(datapath = "path/to/BLUPS.xls")

}
