% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_barplot_checkdiff}
\alias{env_barplot_checkdiff}
\title{env_barplot_checkdiff}
\usage{
env_barplot_checkdiff(import_data, checks)
}
\arguments{
\item{import_data}{selection index dataframe to calculate check difference}

\item{checks}{a list of checks in our dataframe to use as a benchmark for our calculation}
}
\value{
a dataframe datatype
}
\description{
env_barplot_checkdiff
}
\examples{
# calculate_sindex_checkmean(dataframe = "sindex.csv", checks_list)
}
