% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_heatmap}
\alias{env_heatmap}
\title{env_correlation}
\usage{
env_heatmap(dataframe)
}
\arguments{
\item{dataframe}{selection index dataframe to calculate check difference}

\item{checks}{a list of checks in our dataframe to use as a benchmark for our calculation}
}
\value{
a dataframe datatype
}
\description{
env_correlation
}
\examples{
# calculate_sindex_checkmean(dataframe = "sindex.csv", checks_list)
}
