% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_superhheat_corr}
\alias{env_superhheat_corr}
\title{env_superheat_corr}
\usage{
env_superhheat_corr(dataframe, checks)
}
\arguments{
\item{dataframe}{selection index dataframe to calculate check difference}

\item{checks}{a list of checks in our dataframe to use as a benchmark for our calculation}
}
\value{
a dataframe datatype
}
\description{
env_superheat_corr
}
\examples{
# calculate_sindex_checkmean(dataframe = "sindex.csv", checks_list)
}
