% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{export_si_from_blups}
\alias{export_si_from_blups}
\title{export_si_from_blups}
\usage{
export_si_from_blups(datapath, ...)
}
\arguments{
\item{datapath}{BLUPS data path to check for sindex sheet}
}
\value{
helps to extract the selection index data in BLUPS file
}
\description{
export_si_from_blups
}
\examples{
# export_si_from_blups(datapath = "path/to/BLUPS.xls")
}
