% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{fix_listOfTop_Traits}
\alias{fix_listOfTop_Traits}
\title{fix_listOfTop_Traits}
\usage{
fix_listOfTop_Traits(original_list, top_traits)
}
\arguments{
\item{top_traits}{}
}
\value{
a list
}
\description{
fix_listOfTop_Traits
}
\examples{
# check_si(datapath = "path/to/BLUPS.xls")

}
