% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{generate_sindex_table_from_csv}
\alias{generate_sindex_table_from_csv}
\title{generate_sindex_data_from_csv}
\usage{
generate_sindex_table_from_csv(dataframe)
}
\arguments{
\item{dataframe}{BLUPS dataframe to extract sindex sheet}
}
\value{
a dataframe to calculate selection index
}
\description{
generate_sindex_data_from_csv
}
\examples{
# generate_sindex_data(dataframe = "dataframe")

}
