% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{generate_sindex_table_from_excel}
\alias{generate_sindex_table_from_excel}
\title{generate_sindex_data}
\usage{
generate_sindex_table_from_excel(datapath)
}
\arguments{
\item{datapath}{BLUPS data path to check for sindex sheet}
}
\value{
a dataframe that includes selection index
}
\description{
generate_sindex_data
}
\examples{
# generate_sindex_data(datapath = "path/to/BLUPS.xls")

}
