% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_checks}
\alias{get_checks}
\title{get_checks}
\usage{
get_checks(accession_list)
}
\arguments{
\item{accession_list}{the list of all the accession in your data}
}
\value{
a list of detected checks in the dataframe
}
\description{
get_checks
}
\examples{
# get_checks(accession_list = dataframe$accession)

}
