% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{get_genotype_by_location_data}
\alias{get_genotype_by_location_data}
\title{get_genotype_by_location_data}
\usage{
get_genotype_by_location_data(datapath, ...)
}
\arguments{
\item{datapath}{the exact path to your blups data}
}
\value{
a stacked dataframe of all the traits and location in one dataframe extracted from excel sheets
}
\description{
get_genotype_by_location_data
}
\examples{
# get_genotype_by_location_data(datapath = "path/to/BLUPS.xls")

}
