% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{linePlot}
\alias{linePlot}
\title{linePlot}
\usage{
linePlot(trait_to_plot, imported_data, checks, accession_range)
}
\arguments{
\item{datapath}{the exact path to your blups data}
}
\value{
a stacked dataframe of all the traits and location in one dataframe
}
\description{
linePlot
}
\examples{
# get_genotype_by_location_data(datapath = "path/to/BLUPS.xls")

}
