% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{multiyears_tidy_data}
\alias{multiyears_tidy_data}
\title{multiyears_tidy_data}
\usage{
multiyears_tidy_data(dataframe)
}
\arguments{
\item{dataframe}{A multiyears data frame in the regular format}
}
\value{
a tidied dataframe for visualization
}
\description{
multiyears_tidy_data
}
\examples{
#multiyears_tidy_data(dataframe)
}
