% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{multiyears_wrangler}
\alias{multiyears_wrangler}
\title{multiyears_wrangler}
\usage{
multiyears_wrangler(dataframe)
}
\arguments{
\item{dataframe}{a dataframe for wrangling multiyears dataframe}
}
\value{
a wrangled dataframe
}
\description{
multiyears_wrangler
}
\examples{
multiyears_wrangler(dataframe)
}
