% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_checks}
\alias{plot_checks}
\title{plot_checks}
\usage{
plot_checks(dataframe, traits)
}
\arguments{
\item{datapath}{the exact path to your blups data}
}
\value{
a stacked dataframe of all the traits and location in one dataframe
}
\description{
plot_checks
}
\examples{
# get_genotype_by_location_data(datapath = "path/to/BLUPS.xls")

}
