% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_finlay_chart}
\alias{plot_finlay_chart}
\title{plot_finlay_chart}
\usage{
plot_finlay_chart(imported_data, trait_to_plot, checks, accession_range)
}
\arguments{
\item{imported_data}{the dataframe used, containing locations and their means}

\item{trait_to_plot}{the selected trait to plot}
}
\value{
a 'finlay' kind of plot showing the performance of
accessions across different environments, scaled by the average
}
\description{
plot_finlay_chart
}
\examples{
# plot_finlay_chart(imported_data = mydata, trait_to_plot = "DYLD")

}
