% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precdata.R
\docType{data}
\name{precdata}
\alias{precdata}
\title{Precipitation data}
\format{
A shape file, with longitude,latitude and layers
}
\source{
\url{IITA: Breeding Program}
}
\usage{
precdata
}
\description{
A dataset containing precipitation data in Nigeria
}
\keyword{datasets}
