% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapedata.R
\docType{data}
\name{shapedata}
\alias{shapedata}
\title{shape data}
\format{
A shape file, with 37 rows and about 7 columns including the statename, capital, multipolygon data
}
\source{
\url{IITA: Breeding Program}
}
\usage{
shapedata
}
\description{
A dataset containing Shape data in Nigeria MAP
}
\keyword{datasets}
