% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdata.R
\docType{data}
\name{tempdata}
\alias{tempdata}
\title{Temperature data}
\format{
A shape file, with longitude,latitude and layers
}
\source{
\url{IITA: Breeding Program}
}
\usage{
tempdata
}
\description{
A dataset containing precipitation data in Nigeria
}
\keyword{datasets}
