% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{wrangle_colnames}
\alias{wrangle_colnames}
\title{generate_sindex_data}
\usage{
wrangle_colnames(data)
}
\arguments{
\item{data}{rename headers of blups dataframe e.g ag_40 to Ago_owu, ib_12 = Ibadan}
}
\value{
a dataframe with the exact correct names of location
}
\description{
generate_sindex_data
}
\examples{
# generate_sindex_data(datapath = "path/to/BLUPS.xls")
}
