% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{wrangle_data}
\alias{wrangle_data}
\title{wrangle_data}
\usage{
wrangle_data(imported_data)
}
\arguments{
\item{imported_data}{check headers that are not spelt correctly and put necessary corrections}
}
\value{
a dataframe with the exact correct names of location
}
\description{
wrangle_data
}
\examples{
# generate_sindex_data(dataframe = "blupscsvdata/exceldata")
# target headers are accession, combined, and trait

}
