% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoy.R
\name{wrangle_data_sindex}
\alias{wrangle_data_sindex}
\title{wrangle_data_sindex}
\usage{
wrangle_data_sindex(imported_data)
}
\arguments{
\item{imported_data}{rename headers of blups dataframe e.g ag_40 to Ago_owu, ib_12 = Ibadan}
}
\value{
a dataframe with the exact correct names of location
}
\description{
wrangle_data_sindex
}
\examples{
# generate_sindex_data(datapath = "path/to/BLUPS.xls")

}
